pyproject.toml
# pyproject.toml
[project]
name = "feed-aggregator-backend"
version = "0.1.0"
description = "Simple RSS Feed Reader"
authors = [
    { name = "Adwait", email = "adwaitkadam1609@gmail.com" }
]


license = { text = "MIT" }
requires-python = ">=3.11"

dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy==2.0.23",
    "alembic==1.12.1",
    "feedparser==6.0.10",
    "pydantic-settings==2.0.3",
    "python-multipart==0.0.6",
    "loguru==0.7.2",
    "beautifulsoup4==4.12.2",
    "requests==2.31.0",
    "python-dotenv==1.0.0",
    "apscheduler==3.10.4",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "email-validator==2.1.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-env>=0.8.0",
    "httpx>=0.24.0",  # For FastAPI test client
    "ruff",
    "pre-commit",
    "ipykernel",
    "mypy",
    "types-requests",
    "types-beautifulsoup4"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "database", "jobs", "models", "processors", "utils"]

